Implement server response processing through iteratees
The implementation is based on the idea of a Broadcast that broadcasts
messages to any number of registered consumers. The consumers are
defined in terms of iteratees, i.e. they reduce the stream of replies to
some return value when they finally terminate.

The implementation is based on Control.Concurrent.Chan, which makes it
unnecessary to handle consumer management explicitly: A consumer
registers an IO action that is supposed to initiate some communication
with an external entity and returns an iteratee to "consume" the
generated replies. Internally, the main communication channel is
duplicated, so that anything delivered will be available on the
consumer's channel, too. The IO action is invoked after the channel is
duplicated, eliminating a possible race condition, where the request is
sent and the reply is received before the channel has been duplicated.

An enumerator is provided for enumerating a Control.Concurrent.Chan, as
well as some iteratees specific to OSC message processing.