Add support for building an OSX application bundle
Libraries and frameworks are not copied yet, because otool reports a
relative path for the libraries qtHaskell links against.

The database is read from the application's data directory by default,
but can be changed from the command line, along with the database file
pattern (which cannot be change from the UI yet).