Use qBrush_nf constructors to fix random crashes
Previously, the application sometimes crashed with the following error
message:

Mescaline: error: a C finalizer called back into Haskell.
   This was previously allowed, but is disallowed in GHC 6.10.2 and
   later.
   To create finalizers that may call back into Haskll, use
   Foreign.Concurrent.newForeignPtr instead of Foreign.newForeignPtr.

This is what David Harley had to say:

"I'm convinced the problem in question is caused by a call to the C++
routine "freeDynamicHandlers" which is called from the finalizers of non
QObject based objects which have virtual functions. This routine should
only be called there when the _nf constructor is not used and handlers
have been set on the object. I know that it seems to occur spuriously
when handlers have not been set on any objects in a program, but in my
experience so far, the use of _nf constructor solves the problem
completely. At the moment even legitimate programs which set handlers on
non _nf/non-QObject objects will fail with the RTS message about
concurrency, but I'm a bit hesitant about using
Foreign.Concurrent.newForeignPointer to fix this. I'd rather find out
what was really going on first. Also to avoid using
Foreign.Concurrent.newForeignPointer, I could just make dynamic handlers
illegal on non-QObject objects that aren't created using a _nf
constructor."