Fix memory leak in Sound.SC3.Server.Connection
Sound.OpenSoundControl.BufferedTransport is flawed in the sense that if
the primary channel (that is being dup'd by fork) is not read from,
memory accumulates without bounds.

Sound.SC3.Server.Connection has been remodeled to explicitly register
listener actions, that are executed when an incoming OSC packet is
received. Race conditions can be avoided by grouping the send action
that produces an asynchronous reply with the registering of the
listener, i.e. the listener is registered before the send action
commands are sent to the server.