Implement a pattern interpreter using a DSL and Hugs
Expressions in the DSL produce an abstract syntax tree, that is written
to a file, read by GHC using a standard Read instance and compiled to a
pattern.

The DSL is based on Oleg Kiselyov's post on sharing to the haskell-cafe
mailing list (http://okmij.org/ftp/Haskell/DSLSharing.hs). Expressions
are constructed in a completely type safe manner and the compiler is
guaranteed not to produce any runtime errors.

Sharing has to be made explicit in the object language via a bind
construct.