Implement bind construct for patterns and use runtime value binding instead of expression sharing
The bind construct from the object language is now transferred to the
implementation (pattern) language, such that the current value of a
pattern bound in a bind expression is cached and accessed in the body of
the bind.

Also get rid of the cursor field in Event and retrieve the value at the
cursor directly from the environment. The step function also modifies
the environment directly.