Normalize features in unit tree to fix kd-Tree distance "bug"
In order for distances in the 2D visualization to coincide with distances in feature space, the different dimensions need to be normalized (e.g. by subtracting the minimum value and scaling by the reciprocal of the maximum feature bandwidth). This only works when the scatter plot is quadratic and thus has the same scale on each axis; otherwise the ration would need to be taken into account, too.